flip the bit to invert dependency and make core API module independent on any java libraries.
