+++ google-cloud-apis/build.gradle
@@ -22,15 +22,10 @@ sourceSets.test.java.srcDirs = ['testSrc']
 sourceSets.test.resources.srcDirs = ['testResources']
 
 intellij {
-    plugins 'Groovy', 'gradle', 'git4idea', 'properties','maven', 'yaml'
 }
 
 dependencies {
     compile(project(':google-cloud-core'))
-    compile(project(':app-engine:java'))
-    compile(project(':app-engine:java:maven'))
-    // TODO revert dependency once Maven module is complete
-    compile(project(':google-cloud-apis:java-apis:maven-apis'))
     compile 'com.google.apis:google-api-services-servicemanagement:v1-rev374-1.22.0'
     compile ('com.google.cloud.tools:appengine-plugins-core:' + toolsLibVersion) {
         exclude group: 'com.google.guava', module: 'guava'
