+++ google-cloud-apis/java-apis/maven-apis/build.gradle
@@ -27,6 +27,9 @@ intellij {
 
 dependencies {
     compile(project(':google-cloud-core'))
+    compile(project(':app-engine:java'))
+    compile(project(':app-engine:java:maven'))
+    compile(project(':google-cloud-apis'))
     compile ('com.google.cloud.tools:appengine-plugins-core:' + toolsLibVersion) {
         exclude group: 'com.google.guava', module: 'guava'
         exclude group: 'org.yaml', module: 'snakeyaml'
@@ -34,6 +37,4 @@ dependencies {
 
     testCompile files("${System.properties['java.home']}/../lib/tools.jar")
     testCompile(project(':common-test-lib'))
-    // TODO: remove once dependency inverts
-    testRuntime(project(':google-cloud-apis'))
 }
