+++ google-cloud-apis/java-apis/maven-apis/resources/META-INF/cloud-apis-java-maven.xml
@@ -25,8 +25,6 @@
     </project-components>
     <extensions defaultExtensionNs="com.intellij">
         <applicationService
-                serviceImplementation="com.google.cloud.tools.intellij.cloudapis.maven.CloudLibrariesService"/>
-        <applicationService
                 serviceImplementation="com.google.cloud.tools.intellij.cloudapis.maven.CloudApiMavenService"/>
         <localInspection language="XML" shortName="DependencyVersionWithBom"
                          bundle="messages.MavenCloudApisBundle"
@@ -47,5 +45,7 @@
     <extensions defaultExtensionNs="com.google.gct.cloudapis">
         <cloudApiUiExtension
                 implementation="com.google.cloud.tools.intellij.cloudapis.maven.MavenCloudApiUiExtension"/>
+        <cloudApiActionDecorator
+                implementation="com.google.cloud.tools.intellij.cloudapis.maven.MavenCloudApiActionDecorator"/>
     </extensions>
 </idea-plugin>
