+++ google-cloud-apis/java-apis/maven-apis/resources/messages/MavenCloudApisBundle.properties
@@ -14,6 +14,11 @@
 # limitations under the License.
 #
 
+cloud.libraries.menu.action.disabled.java8.text=Add Cloud libraries (requires Java 8)...
+cloud.libraries.menu.action.disabled.maven.text=Add Cloud libraries (requires Maven)...
+cloud.libraries.menu.action.maven.required.description=Only Maven based projects are supported at this time
+cloud.libraries.menu.action.gae.java8.required.description=These libraries are not compatible with App Engine Standard Java 7, use the App Engine SDK instead
+
 cloud.libraries.bom.inspection.group.name=Google Cloud Java Library BOM Inspections
 cloud.libraries.version.with.bom.inspection.display.name=Dependency Version with BOM Inspection
 cloud.libraries.version.with.bom.inspection.description=Inspection that checks for the presence of version tags in google cloud java Maven dependencies when a BOM has been imported.
