+++ google-cloud-apis/resources/META-INF/google-cloud-apis.xml
@@ -22,11 +22,15 @@
                         interface="com.google.cloud.tools.intellij.cloudapis.CloudApiRunConfigurationProvider"/>
         <extensionPoint qualifiedName="com.google.gct.cloudapis.cloudApiUiExtension"
                         interface="com.google.cloud.tools.intellij.cloudapis.CloudApiUiExtension"/>
+        <extensionPoint qualifiedName="com.google.gct.cloudapis.cloudApiActionDecorator"
+                        interface="com.google.cloud.tools.intellij.cloudapis.CloudApiActionDecoratorExtension"/>
     </extensionPoints>
     <extensions defaultExtensionNs="com.intellij">
         <applicationService
                 serviceInterface="com.google.cloud.tools.intellij.cloudapis.CloudApiUiPresenter"
                 serviceImplementation="com.google.cloud.tools.intellij.cloudapis.DefaultCloudApiUiPresenter"/>
+        <applicationService
+                serviceImplementation="com.google.cloud.tools.intellij.cloudapis.CloudLibrariesService"/>
     </extensions>
     <actions>
         <group id="GoogleCloudTools.cloudapis">
