+++ google-cloud-apis/resources/messages/CloudApisBundle.properties
@@ -17,11 +17,7 @@
 cloud.libraries.dialog.title=Add Google Cloud Libraries
 cloud.libraries.documentation.link=Documentation
 cloud.libraries.menu.action.active.text=Add Cloud libraries...
-cloud.libraries.menu.action.disabled.java8.text=Add Cloud libraries (requires Java 8)...
-cloud.libraries.menu.action.disabled.maven.text=Add Cloud libraries (requires Maven)...
 cloud.libraries.menu.action.description=Select the Cloud libraries to download and include in your application
-cloud.libraries.menu.action.maven.required.description=Only Maven based projects are supported at this time
-cloud.libraries.menu.action.gae.java8.required.description=These libraries are not compatible with App Engine Standard Java 7, use the App Engine SDK instead
 cloud.libraries.module.selector.label=Add Google Cloud Library to Module:
 cloud.libraries.ok.button.text=Add Client Library
 cloud.apis.enable.checkbox.text=Enable the API on the Google Cloud Platform
