+++ google-cloud-apis/src/com/google/cloud/tools/intellij/cloudapis/AddCloudLibrariesAction.java
@@ -17,19 +17,9 @@
 package com.google.cloud.tools.intellij.cloudapis;
 
 import com.google.cloud.tools.intellij.GoogleCloudCoreIcons;
-import com.google.cloud.tools.intellij.appengine.java.facet.standard.AppEngineStandardFacet;
-import com.google.cloud.tools.intellij.appengine.java.facet.standard.AppEngineStandardFacetType;
-import com.google.cloud.tools.intellij.appengine.java.maven.project.MavenProjectService;
-import com.google.cloud.tools.intellij.appengine.java.project.AppEngineProjectService;
-import com.intellij.facet.FacetManager;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.module.Module;
-import com.intellij.openapi.module.ModuleManager;
 import com.intellij.openapi.project.DumbAwareAction;
-import java.util.Set;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.apache.commons.lang3.JavaVersion;
 
 /**
  * The action in the Google Cloud Tools menu group that opens the wizard to add client libraries to
@@ -47,52 +37,8 @@ public final class AddCloudLibrariesAction extends DumbAwareAction {
   @Override
   @SuppressWarnings("MissingCasesInEnumSwitch")
   public void update(AnActionEvent e) {
-    boolean addLibrariesEnabled = false;
-    if (e.getProject() != null) {
-      Set<CloudLibrariesModuleSupportType> moduleSupportTypes =
-          Stream.of(ModuleManager.getInstance(e.getProject()).getModules())
-              .map(this::checkModuleForAddCloudLibraries)
-              .collect(Collectors.toSet());
-
-      addLibrariesEnabled = moduleSupportTypes.contains(CloudLibrariesModuleSupportType.SUPPORTED);
-      if (!addLibrariesEnabled) {
-        // update message to hint what is missing.
-        for (CloudLibrariesModuleSupportType supportType : moduleSupportTypes) {
-          switch (supportType) {
-            case MAVEN_REQUIRED:
-              e.getPresentation()
-                  .setDescription(
-                      GoogleCloudApisMessageBundle.message(
-                          "cloud.libraries.menu.action.maven.required.description"));
-              e.getPresentation()
-                  .setText(
-                      GoogleCloudApisMessageBundle.message(
-                          "cloud.libraries.menu.action.disabled.maven.text"));
-              break;
-            case APPENGINE_JAVA8_REQUIRED:
-              e.getPresentation()
-                  .setDescription(
-                      GoogleCloudApisMessageBundle.message(
-                          "cloud.libraries.menu.action.gae.java8.required.description"));
-              e.getPresentation()
-                  .setText(
-                      GoogleCloudApisMessageBundle.message(
-                          "cloud.libraries.menu.action.disabled.java8.text"));
-              break;
-          }
-        }
-      } else {
-        // standard message for a supported module action.
-        e.getPresentation()
-            .setDescription(
-                GoogleCloudApisMessageBundle.message("cloud.libraries.menu.action.description"));
-        e.getPresentation()
-            .setText(
-                GoogleCloudApisMessageBundle.message("cloud.libraries.menu.action.active.text"));
-      }
-    }
-
-    e.getPresentation().setEnabled(addLibrariesEnabled);
+    Stream.of(CloudApiActionDecoratorExtension.EP_NAME.getExtensions())
+        .forEach(decorator -> decorator.decorate(e));
   }
 
   @Override
@@ -110,25 +56,5 @@ public final class AddCloudLibrariesAction extends DumbAwareAction {
         }
       }
     }
-  }
-
-  private enum CloudLibrariesModuleSupportType {
-    SUPPORTED,
-    MAVEN_REQUIRED,
-    APPENGINE_JAVA8_REQUIRED
-  }
-
-  private CloudLibrariesModuleSupportType checkModuleForAddCloudLibraries(Module module) {
-    // AppEngine Standard + Java 7 are not supported for GCP Libraries
-    if (AppEngineProjectService.getInstance().hasAppEngineStandardFacet(module)) {
-      AppEngineStandardFacet appEngineStandardFacet =
-          FacetManager.getInstance(module).getFacetByType(AppEngineStandardFacetType.ID);
-      if (!appEngineStandardFacet.getRuntimeJavaVersion().atLeast(JavaVersion.JAVA_1_8))
-        return CloudLibrariesModuleSupportType.APPENGINE_JAVA8_REQUIRED;
-    }
-
-    return MavenProjectService.getInstance().isMavenModule(module)
-        ? CloudLibrariesModuleSupportType.SUPPORTED
-        : CloudLibrariesModuleSupportType.MAVEN_REQUIRED;
   }
 }
