+++ google-cloud-apis/src/com/google/cloud/tools/intellij/cloudapis/AddCloudLibrariesDialog.java
@@ -17,7 +17,6 @@
 package com.google.cloud.tools.intellij.cloudapis;
 
 import com.google.api.services.iam.v1.model.Role;
-import com.google.cloud.tools.intellij.cloudapis.maven.CloudLibrariesService;
 import com.google.cloud.tools.intellij.project.CloudProject;
 import com.google.cloud.tools.libraries.json.CloudLibrary;
 import com.google.common.collect.Sets;
