+++ google-cloud-apis/src/com/google/cloud/tools/intellij/cloudapis/CloudLibrariesService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.cloudapis.maven;
+package com.google.cloud.tools.intellij.cloudapis;
 
 import com.google.cloud.tools.libraries.CloudLibraries;
 import com.google.cloud.tools.libraries.json.CloudLibrary;
